+++ src/context/sidebar.tsx
@@ -5,17 +5,22 @@ import { createContext, useContext, useMemo, useState } from "react"
 interface TreeSidebarContextValues{
      open: boolean,
      setOpen: SetStateType<boolean>,
-     isMobile: boolean
+     isMobile: boolean,
+     closeMobileSidebar: () => void
 }
 const TreeSidebarContext = createContext<TreeSidebarContextValues | null>(null)
 
 export function TreeSidebarProvider({ children }: { children: React.ReactNode }){
-     const [open, setOpen] = useState(false);
-     const isMobile = useIsMobile()
+     const [open, setOpen] = useState(true);
+     const isMobile = useIsMobile();
+     const closeMobileSidebar = () => {
+          if (isMobile) setOpen(false)
+     }
      const values: TreeSidebarContextValues = useMemo(()=>({
           open,
           setOpen,
-          isMobile
+          isMobile,
+          closeMobileSidebar
      }),[open, isMobile])
      return (
           <TreeSidebarContext.Provider value={values}>